Rename OpenRPC classes and create new module
Renamed OpenRPC-related classes and packages for better organization and clarity. Introduced a new Maven module, `open-rpc-schema`, to house the schema-specific code and dependencies.
