+++ open-rpc-schema/src/main/java/io/quarkiverse/jsonrpc/openrpc/schema/License.java
@@ -1,4 +1,4 @@
-package io.quarkiverse.jsonrpc.deployment.openrpc.spec;
+package io.quarkiverse.jsonrpc.openrpc.schema;
 
 public record License(
         String name,
